feat: wire logic on  settings page
