+++ assets/settings.html
@@ -1,6 +1,7 @@
 <html>
     <head>
         <title>TweetGPT Settings</title>
+        <meta charset="UTF-8">
         <link rel="preconnect" href="https://fonts.googleapis.com">
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
         <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap" rel="stylesheet">
