+++ lib/settings.js
@@ -443,7 +443,7 @@ exports.unstable_renderSubtreeIntoContainer=function(a,b,c,d){if(!pl(c))throw Er
 
 /***/ }),
 
-/***/ 265:
+/***/ 745:
 /***/ ((__unused_webpack_module, exports, __webpack_require__) => {
 
 "use strict";
@@ -740,7 +740,7 @@ var __webpack_exports__ = {};
 // EXTERNAL MODULE: ./node_modules/react/index.js
 var react = __webpack_require__(294);
 // EXTERNAL MODULE: ./node_modules/react-dom/client.js
-var client = __webpack_require__(265);
+var client = __webpack_require__(745);
 // EXTERNAL MODULE: ./node_modules/react-is/index.js
 var react_is = __webpack_require__(864);
 // EXTERNAL MODULE: ./node_modules/shallowequal/index.js
@@ -1472,26 +1472,6 @@ const Container = styled_components_browser_esm.div `
     display: flex;
     flex-direction: column;
 `;
-const H1 = styled_components_browser_esm.h1 `
-    font-size: 25px;
-    font-weight: bold;
-    color: #181C21;
-    margin: 0;
-`;
-const H2 = styled_components_browser_esm.h2 `
-    font-size: 18px;
-    font-weight: 600;
-    color: #181C21;
-    margin: 0;
-`;
-const Subtitle = styled_components_browser_esm.span `
-    font-size: 15px;
-    color: #566370;
-
-    a {
-        color: #566370;
-    }
-`;
 const Logo = styled_components_browser_esm.img `
     width: 80px;
     height: 80px;
@@ -1505,26 +1485,8 @@ const SettingsBlock = styled_components_browser_esm.div `
     align-items: flex-start;
     flex-direction: column;
 `;
-const TextInput = styled_components_browser_esm.input `
-    box-sizing: border-box;
-    width: 100%;
-    height: 40px;
-    padding: 0 10px;
-    background: #FFFFFF;
-    border: 1px solid #D1D9DD;
-    color: #181C21;
-    font-size: 15px;
-    border-radius: 5px;
 
-    ::placeholder {
-        color: #D1D9DD;
-    }
-`;
-const Space = styled_components_browser_esm.div((props) => `
-    height: ${props.height}px;
-`);
-
-;// CONCATENATED MODULE: ./src/settings/components/CheckBox.tsx
+;// CONCATENATED MODULE: ./src/settings/components/checkbox.tsx
 
 
 const icons = {
@@ -1533,13 +1495,13 @@ const icons = {
 };
 const Checkbox = ({ value, label, onChange }) => {
     const src = value && icons.checked || icons.uncheked;
-    return (react.createElement(CheckBox_Container, null,
+    return (react.createElement(checkbox_Container, { onClick: () => onChange(!value) },
         react.createElement("div", { className: 'iconContainer' },
             react.createElement(Icon, { src: src }),
-            react.createElement(CheckboxSquare, { type: "checkbox", name: label, checked: value, onClick: () => onChange(!value) })),
+            react.createElement(CheckboxSquare, { type: "checkbox", name: label, checked: value })),
         react.createElement(Label, { htmlFor: label }, label)));
 };
-const CheckBox_Container = styled_components_browser_esm.div `
+const checkbox_Container = styled_components_browser_esm.div `
     cursor: pointer;
 
     .iconContainer {
@@ -1574,6 +1536,7 @@ const Label = styled_components_browser_esm.label `
     font-size: 15px;
     cursor: pointer;
     color: #181C21;
+    user-select: none;
 `;
 
 ;// CONCATENATED MODULE: ./src/settings/components/selector.tsx
@@ -1602,7 +1565,284 @@ const Select = styled_components_browser_esm.select `
     box-sizing: border-box;
 `;
 
+;// CONCATENATED MODULE: ./src/settings/components/text_input.tsx
+
+const TextInput = styled_components_browser_esm.input `
+    box-sizing: border-box;
+    width: 100%;
+    height: 40px;
+    padding: 0 10px;
+    background: #FFFFFF;
+    border: 1px solid #D1D9DD;
+    color: #181C21;
+    font-size: 15px;
+    border-radius: 5px;
+
+    ::placeholder {
+        color: #D1D9DD;
+    }
+`;
+
+;// CONCATENATED MODULE: ./src/settings/components/text.tsx
+
+const H1 = styled_components_browser_esm.h1 `
+    font-size: 25px;
+    font-weight: bold;
+    color: #181C21;
+    margin: 0;
+`;
+const H2 = styled_components_browser_esm.h2 `
+    font-size: 18px;
+    font-weight: 600;
+    color: #181C21;
+    margin: 0;
+`;
+const Subtitle = styled_components_browser_esm.span `
+    font-size: 15px;
+    color: #566370;
+
+    a {
+        color: #566370;
+    }
+`;
+
+;// CONCATENATED MODULE: ./src/settings/components/space.tsx
+
+const Space = styled_components_browser_esm.div((props) => `
+    height: ${props.height}px;
+`);
+
+;// CONCATENATED MODULE: ./src/background/chat_gpt_client/locales.ts
+const locales = {
+    "af-ZA": [
+        "Afrikaans",
+        "Afrikaans"
+    ],
+    "ar": [
+        "العربية",
+        "Arabic"
+    ],
+    "bg-BG": [
+        "Български",
+        "Bulgarian"
+    ],
+    "ca-AD": [
+        "Català",
+        "Catalan"
+    ],
+    "cs-CZ": [
+        "Čeština",
+        "Czech"
+    ],
+    "cy-GB": [
+        "Cymraeg",
+        "Welsh"
+    ],
+    "da-DK": [
+        "Dansk",
+        "Danish"
+    ],
+    "de-AT": [
+        "Deutsch (Österreich)",
+        "German (Austria)"
+    ],
+    "de-CH": [
+        "Deutsch (Schweiz)",
+        "German (Switzerland)"
+    ],
+    "de-DE": [
+        "Deutsch (Deutschland)",
+        "German (Germany)"
+    ],
+    "el-GR": [
+        "Ελληνικά",
+        "Greek"
+    ],
+    "en-GB": [
+        "English (UK)",
+        "English (UK)"
+    ],
+    "en-US": [
+        "English (US)",
+        "English (US)"
+    ],
+    "es-CL": [
+        "Español (Chile)",
+        "Spanish (Chile)"
+    ],
+    "es-ES": [
+        "Español (España)",
+        "Spanish (Spain)"
+    ],
+    "es-MX": [
+        "Español (México)",
+        "Spanish (Mexico)"
+    ],
+    "et-EE": [
+        "Eesti keel",
+        "Estonian"
+    ],
+    "eu": [
+        "Euskara",
+        "Basque"
+    ],
+    "fa-IR": [
+        "فارسی",
+        "Persian"
+    ],
+    "fi-FI": [
+        "Suomi",
+        "Finnish"
+    ],
+    "fr-CA": [
+        "Français (Canada)",
+        "French (Canada)"
+    ],
+    "fr-FR": [
+        "Français (France)",
+        "French (France)"
+    ],
+    "he-IL": [
+        "עברית",
+        "Hebrew"
+    ],
+    "hi-IN": [
+        "हिंदी",
+        "Hindi"
+    ],
+    "hr-HR": [
+        "Hrvatski",
+        "Croatian"
+    ],
+    "hu-HU": [
+        "Magyar",
+        "Hungarian"
+    ],
+    "id-ID": [
+        "Bahasa Indonesia",
+        "Indonesian"
+    ],
+    "is-IS": [
+        "Íslenska",
+        "Icelandic"
+    ],
+    "it-IT": [
+        "Italiano",
+        "Italian"
+    ],
+    "ja-JP": [
+        "日本語",
+        "Japanese"
+    ],
+    "km-KH": [
+        "ភាសាខ្មែរ",
+        "Khmer"
+    ],
+    "ko-KR": [
+        "한국어",
+        "Korean"
+    ],
+    "la": [
+        "Latina",
+        "Latin"
+    ],
+    "lt-LT": [
+        "Lietuvių kalba",
+        "Lithuanian"
+    ],
+    "lv-LV": [
+        "Latviešu",
+        "Latvian"
+    ],
+    "mn-MN": [
+        "Монгол",
+        "Mongolian"
+    ],
+    "nb-NO": [
+        "Norsk bokmål",
+        "Norwegian (Bokmål)"
+    ],
+    "nl-NL": [
+        "Nederlands",
+        "Dutch"
+    ],
+    "nn-NO": [
+        "Norsk nynorsk",
+        "Norwegian (Nynorsk)"
+    ],
+    "pl-PL": [
+        "Polski",
+        "Polish"
+    ],
+    "pt-BR": [
+        "Português (Brasil)",
+        "Portuguese (Brazil)"
+    ],
+    "pt-PT": [
+        "Português (Portugal)",
+        "Portuguese (Portugal)"
+    ],
+    "ro-RO": [
+        "Română",
+        "Romanian"
+    ],
+    "ru-RU": [
+        "Русский",
+        "Russian"
+    ],
+    "sk-SK": [
+        "Slovenčina",
+        "Slovak"
+    ],
+    "sl-SI": [
+        "Slovenščina",
+        "Slovenian"
+    ],
+    "sr-RS": [
+        "Српски / Srpski",
+        "Serbian"
+    ],
+    "sv-SE": [
+        "Svenska",
+        "Swedish"
+    ],
+    "th-TH": [
+        "ไทย",
+        "Thai"
+    ],
+    "tr-TR": [
+        "Türkçe",
+        "Turkish"
+    ],
+    "uk-UA": [
+        "Українська",
+        "Ukrainian"
+    ],
+    "ur-PK": [
+        "اردو",
+        "Urdu"
+    ],
+    "vi-VN": [
+        "Tiếng Việt",
+        "Vietnamese"
+    ],
+    "zh-CN": [
+        "中文 (中国大陆)",
+        "Chinese (PRC)"
+    ],
+    "zh-TW": [
+        "中文 (台灣)",
+        "Chinese (Taiwan)"
+    ]
+};
+const defaultLocale = 'en-US';
+
 ;// CONCATENATED MODULE: ./src/settings/settings_screen.tsx
+
+
+
+
+
 
 
 
@@ -1610,29 +1850,78 @@ const Select = styled_components_browser_esm.select `
 
 const SettingsScreen = () => {
     const logoPath = (0,react.useMemo)(() => chrome.runtime.getURL('icons/128.png'), []);
+    const [openAIToken, setOpenAIToken] = (0,react.useState)('');
+    const [selectedLanguage, setLanguage] = (0,react.useState)(defaultLocale);
+    const [isAddSignature, setAddSignature] = (0,react.useState)(true);
+    const [isAddTopicForReplies, setTopicForReplies] = (0,react.useState)(false);
+    const changeOpenAIToken = (0,react.useCallback)(async (newToken) => {
+        await chrome.storage.local.set({ openAIToken: newToken });
+        setOpenAIToken(newToken);
+    }, []);
+    const changeSignature = (0,react.useCallback)(async (isNewAddSignature) => {
+        console.log('>>>', isNewAddSignature);
+        await chrome.storage.local.set({ isAddSignature: isNewAddSignature });
+        setAddSignature(isNewAddSignature);
+    }, []);
+    const changeTopic = (0,react.useCallback)(async (isAddTopicForReplies) => {
+        await chrome.storage.local.set({ isAddTopicForReplies });
+        setTopicForReplies(isAddTopicForReplies);
+    }, []);
+    const changeLanguage = (0,react.useCallback)(async (language) => {
+        await chrome.storage.local.set({ language });
+        setLanguage(language);
+    }, []);
+    (0,react.useEffect)(() => {
+        chrome.storage.local.get(['language', 'isAddSignature', 'isAddTopicForReplies', 'openAIToken']).then((state) => {
+            if (state['language']) {
+                setLanguage(state['language']);
+            }
+            if (state['isAddSignature'] != null) {
+                setAddSignature(state['isAddSignature']);
+            }
+            if (state['isAddTopicForReplies'] != null) {
+                setTopicForReplies(state['isAddTopicForReplies']);
+            }
+            if (state['openAIToken']) {
+                setOpenAIToken(state['openAIToken']);
+            }
+        });
+    }, []);
+    const languageOptions = (0,react.useMemo)(() => {
+        const codes = Object.keys(locales);
+        return codes.map(code => ({
+            value: code,
+            label: locales[code][0]
+        }));
+    }, []);
+    const warning = (0,react.useMemo)(() => {
+        return openAIToken.length === 0 ? '⚠️' : '';
+    }, [openAIToken === '']);
     return (react.createElement(Body, null,
         react.createElement(Logo, { src: logoPath }),
         react.createElement(Container, null,
             react.createElement(H1, null, "TweetGPT Settings"),
             react.createElement(SettingsBlock, null,
-                react.createElement(H2, null, "OpenAI API Token \u26A0\uFE0F"),
+                react.createElement(H2, null,
+                    "OpenAI API Token ",
+                    warning),
                 react.createElement(Space, { height: 3 }),
                 react.createElement(Subtitle, null,
                     "Required. You can find your Secret API key in your ",
                     react.createElement("a", { href: "https://platform.openai.com/account/api-keys" }, "OpenAI User settings"),
                     "."),
                 react.createElement(Space, { height: 10 }),
-                react.createElement(TextInput, { type: "text", placeholder: 'Put your secret API token here' })),
+                react.createElement(TextInput, { type: "text", placeholder: 'Put your secret API token here', value: openAIToken, onChange: (e) => changeOpenAIToken(e.target.value) })),
             react.createElement(SettingsBlock, null,
                 react.createElement(H2, null, "Text Generation"),
                 react.createElement(Space, { height: 10 }),
-                react.createElement(Checkbox, { value: true, onChange: () => { }, label: 'Add TweetGPT Signature' }),
+                react.createElement(Checkbox, { value: isAddSignature, onChange: changeSignature, label: 'Add TweetGPT Signature' }),
                 react.createElement(Space, { height: 10 }),
-                react.createElement(Checkbox, { value: false, onChange: () => { }, label: 'Ask for tweet topic in replies' })),
+                react.createElement(Checkbox, { value: isAddTopicForReplies, onChange: changeTopic, label: 'Ask for tweet topic in replies' })),
             react.createElement(SettingsBlock, null,
                 react.createElement(H2, null, "Language"),
                 react.createElement(Space, { height: 10 }),
-                react.createElement(Selector, { value: 'en', onChange: () => { }, options: [{ value: 'en', label: 'English' }, { value: 'es', label: 'Spanish' }] })))));
+                react.createElement(Selector, { value: selectedLanguage, onChange: changeLanguage, options: languageOptions })))));
 };
 
 ;// CONCATENATED MODULE: ./src/settings/settings.tsx
