--- src/popup/popup.html
@@ -1,34 +1,0 @@
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <style>
-            .checkboxContainer {
-                display:flex;
-                flex-direction: row;
-                align-items: center; 
-                justify-content: flex-start;
-                margin-bottom:10px;
-            }
-        </style>
-    </head>
-    <body style="width:300px">
-        <div class="checkboxContainer">
-            <label for="language">Language&nbsp;&nbsp;&nbsp;</label>
-            <select name="language" id="language">
-            </select>
-        </div>
-        <div class="checkboxContainer">
-            <input type="checkbox" id="isAddSignature" name="isAddSignature"/>
-            <label for="isAddSignature">Add tweetGPT signature</label>
-        </div>
-        <div class="checkboxContainer">
-            <input type="checkbox" id="isAddTopicForReplies" name="isAddTopicForReplies"/>
-            <label for="isAddTopicForReplies">Ask tweet topic for replies</label>
-        </div>
-        <div class="checkboxContainer">
-            <input type="checkbox" id="isRandomType" name="isRandomType" />
-            <label for="isRandomType">Use random tweet types</label>
-        </div>
-    </body>
-    <script src="../../lib/popup.js"></script>
-</html>
