+++ src/settings/components/space.tsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const Space = styled.div((props: { height: number }) => `
+    height: ${props.height}px;
+`);
