+++ src/settings/components/text.tsx
@@ -1,0 +1,24 @@
+import styled from "styled-components";
+
+export const H1 = styled.h1`
+    font-size: 25px;
+    font-weight: bold;
+    color: #181C21;
+    margin: 0;
+`;
+
+export const H2 = styled.h2`
+    font-size: 18px;
+    font-weight: 600;
+    color: #181C21;
+    margin: 0;
+`
+
+export const Subtitle = styled.span`
+    font-size: 15px;
+    color: #566370;
+
+    a {
+        color: #566370;
+    }
+`
