+++ src/settings/components/text_input.tsx
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const TextInput = styled.input`
+    box-sizing: border-box;
+    width: 100%;
+    height: 40px;
+    padding: 0 10px;
+    background: #FFFFFF;
+    border: 1px solid #D1D9DD;
+    color: #181C21;
+    font-size: 15px;
+    border-radius: 5px;
+
+    ::placeholder {
+        color: #D1D9DD;
+    }
+`;
